Fix NameAddrHeader display_name handling (#573)
* Fixes broken quoted_string_clean and display_name parser rules

This fixes 2 problems with the quoted_string_clean rule:
1. The substring operation (`input.substring(pos-1, offset+1);`) to
remove outer quotes returns incorrect results for strings that
begin with SWS
2. It does not completely unescape the string. The outer DQUOTES are
removed but quoted_pair sequences are not unescaped.

The display_name rule is changed in the following ways:
1. quoted_string is changed to quoted_string_clean so that
display_name is properly unescaped when it matches the quoted_string
form of the rule. trimming and unescaping is no longer the
responsibility of display_name
2. logic has been added to collapse arrays of tokens into a string
when the rule matches the `token ( LWS token )*` form

It is worth mentioning that LWS whitespace occuring in the token form
of display_name is now collapsed into a single SP where as before
it was left intact. This is compliant with the spec. (RFC3261 25)

Additionally, a link to the RFC errata documenting the display_name
rule correction has been included to supplement the 'previous rule is
corrected from RFC3261' comment.

* NameAddrHeader: cleans up display_name member variable logic

Force _display_name private member assignment in the constructor to
use the setter function. This removes duplicated logic for dealing
with a display_name of integer 0 in the toString() method

* NameAddrHeader: properly escape display_name in toString()

* SIPMessage: construct To and From headers programmatically

Construct To and From headers by instantiating NameAddrHeader objects
with the appropriate parameters and calling
NameAddrHeader.toString().

The previous logic of creating To and From headers via string
manipulation and then parsing them into NameAddrHeader objects was
clumsy and did not handle display_name with quoted pairs

* Config: removes display_name validation

display_name validation is no longer necessary since proper escaping
and unescaping of display_name has been implemented

* Tests: improves NameAddr parsing tests

Improves parse NameAddr test by adding various combinations of
characters that need to be escaped/unescaped to display_name

Adds test for display_name made up of tokens

Adds test for NameAddr with no space between DQUOTE and LAQUOT

Adds test for NameAddr with no display_name